Refactoring the BFS suite.
Redesigned the pattern by which custom and internal labels are managed within
the graph library. The new design separates the implementation of the algorithm
from the structures that manage the mapping. Internal mappings are explicitly
created in specializations.

Refactored the BFS suite so that it's not a 1000 line file.

Updated label and coloring information accordingly.


git-svn-id: http://origin.googlecode.com/svn/trunk@69 3dbf15f1-5c06-4b2d-4fb8-2ed049db4e7e